test-scenarios/07.1-existing-test-updated-with-failure.html

<!doctype html>
<title>Existing passing test is updated with failure</title>
<!-- In particular, valid non-broken schemes that are nevertheless not going to work -->
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
test(function() {
    assert_true(false);
}, "Passing test"); // This test is now failing after a sync from web-platform-tests
</script>
